ntools_mesh

Struct Voxel

pub struct Voxel {
    pub index: usize,
    pub result: f64,
    pub error: f64,
}

Representation of a single voxel in the mesh

The global index of the voxel is included to maintain consistency between output Format variants. Parsing line-by-line rather than trying to load an entire file into memory leaves the voxels in an inconsistent order for several formats.

Memory usage

The minimum information required would be just the result and uncertainty (16 Bytes), and the maximum would include the cooridinates and energy/time groups (80 Bytes).

The current implementation is a compromise between the two at 24 Bytes, and all other values may be derived from the Mesh given the voxel index.

Several operators are implemented for convenience where it makes sense, including Addition (+, +=), Subtraction (-, -=), Multiplication (*, *=), and Division (/, /=).

In all cases, the LHS index is taken, and the RHS may be either another Voxel or anything that can be converted into an f64 primitive.

Fields
index: usize

Global voxel index

result: f64

Tallied voxel result

error: f64

Relative error on result

Implementations

impl Voxel

pub fn absolute_error(&self) -> f64

Returns the absolute error for the voxel

Example:

 let voxel = Voxel {
     result: 50.0,
     error: 0.10,
     ..Default::default()
 };

 /// 10% relative error => 50.0 +/-5.0
 assert_eq!(voxel.absolute_error(), 5.0);

pub fn relative_error(&self) -> f64

Returns the relative error for the voxel

The MCNP mesh results are output and stored as the relative uncertainty anyway. However, having both absolute_error() and relative_error() methods makes intent explicit.

For example:

let voxel = Voxel {
    result: 50.0,
    error: 0.10,
    ..Default::default()
};
/// 10% relative error => 50.0 +/-5.0
assert_eq!(voxel.relative_error(), 0.1);

pub fn powf(self, value: impl Into<f64>) -> Voxel

Raise the voxel to some power

 let voxel = Voxel {
     result: 10.0,
     error: 0.10,
     ..Default::default()
 };

 /// 10% relative error => 50.0 +/-5.0
 assert_eq!(voxel.powf(2.0).result, 100.0);
 assert_eq!(voxel.powf(2.0).error, 2.0);